pyproject.toml
[tool.black]
line-length = 119

[tool.isort]
line_length = 119
profile = "black"

[tool.mypy]
plugins = ["pydantic.mypy", "strawberry.ext.mypy_plugin"]

[tool.ruff]
line-length = 119
target-version = "py311"

[tool.flake8]
max-line-length = 119
max-complexity = 10
extend-ignore = ["E203", "W605"]

[tool.coverage]
[tool.coverage.run]
omit = ["tests/*", "*conftest.py"]

[tool.pytest.ini_options]
addopts = "--cov-report term-missing --strict-markers"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.poetry]
name = "jelly"
version = "0.1.0"
description = "A playground for FastAPI, GraphQL and Elasticsearch"
license = "MIT"
authors = ["Evgeni Sautin <evgeni.sautin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "0.95.1"
gunicorn = "20.1.0"
uvicorn = { extras = ["standard"], version = "0.22.0" }
strawberry-graphql = { extras = ["fastapi"], version = "0.176.3" }
elasticsearch = {extras = ["async"], version = "^8.7.0"}
motor = "3.1.2"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.7"}

[tool.poetry.group.dev.dependencies]
black = "23.3.0"
isort = "5.12.0"
flake8 = "6.0.0"
flake8-pyproject = "1.2.3"
mypy = "1.1.1"
ruff = "0.0.254"
pre-commit = "3.3.1"
pytest = "7.3.1"
pytest-asyncio = "0.21.0"
pytest-cov = "4.0.0"
pytest-xdist = {extras = ["psutil"], version = "3.2.1"}
httpx = "0.24.0"
faker = "18.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
